Clases datos eventos

+++ app/src/main/java/mx/itesm/aa/atencionatizapan/model/LluviaData.kt
@@ -1,8 +1,8 @@
 package mx.itesm.aa.atencionatizapan.model
 
 data class LluviaData(
-    val humedad: Int,
-    val volumen: Int,
-    val viento: Int,
-    val temperatura: Int
+    val humedad: Float,
+    val volumen: Float,
+    val viento: Float,
+    val temperatura: Float
 )
